requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
pytz==2025.2
pyzbar==0.1.9
sqlparse==0.5.4
tzdata==2025.2
